Moving TreeView icon loading into separate class
