+++ SystemFinder/Abstractions/View/ITreeViewIconLoader.cs
@@ -1,0 +1,7 @@
+﻿namespace SystemFinder.Abstractions.View
+{
+    public interface ITreeViewIconLoader
+    {
+        ImageList LoadTreeViewIcons();
+    }
+}
